Use descriptor to filter fields.
In the original draft of this, if I recall correctly, @awalter-bnl and I
both agreed that we should inspect that array's actual shape to decide
whether it was suitable for writing as TIFF. In a better world, we would
rely on the EventDescriptor for this, but at that time it was known that
many EventDescriptor were not correct and we didn't want suitcase-tiff
to be blocked by a database migration to fix them.

Now, databroker 1.0 gives us a hook to patch erroneous EventDescriptors
using "transforms", and it also gives us new urgency to ensure they
are correct. (If not, dask explodes.) It might be time to rely on them
being correct here as well.

The immediately motivation for this issue is that suitcase-tiff blows up
on strings, which it tries to cast to numpy arrays. Much safer to check
the dtype and shape in the descriptor.