FIX: deal with descriptors that lie about the shape
If the descriptors reports a 3D shape (e.g. (1, 2048, 2048)) and the data
comes out of the filler with the leading dimension dropped, the tiff series
serializer will try to write out each line of the array as its own tiff.

This is because we determined if we needed to "up cast" a 2D array to a 3D
array by consulting the descriptor and then iterate over the data assuming it
is definitely shaped as (frame, rows, cols).

This adds a warning we can make an error in the future.